refactor: 가족 설정 페이지 팝업 정리
가족 설정 화면에서 사용되는 팝업들을 한 파일 안에 정리했습니다.
