+++ app/src/main/java/io/familymoments/app/core/component/popup/RemoveFamilyMemberPopup.kt
@@ -1,4 +1,4 @@
-package io.familymoments.app.feature.removemember
+package io.familymoments.app.core.component.popup
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
