+++ app/src/main/java/io/familymoments/app/feature/updatecycle/screen/UpdateCycleScreen.kt
@@ -5,7 +5,6 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -23,7 +22,8 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import io.familymoments.app.R
 import io.familymoments.app.core.component.FMButton
 import io.familymoments.app.core.component.UploadCycleDropdownMenu
-import io.familymoments.app.core.component.popup.CompletePopUp
+import io.familymoments.app.core.component.popup.FamilyPermissionPopup
+import io.familymoments.app.core.component.popup.FamilySettingCompletePopup
 import io.familymoments.app.core.theme.AppColors
 import io.familymoments.app.core.theme.AppTypography
 import io.familymoments.app.feature.creatingfamily.UploadCycle
@@ -44,28 +44,8 @@ fun UpdateCycleScreen(
     LaunchedEffect(uiState.isOwner) { showPermissionPopup = !uiState.isOwner }
     LaunchedEffect(uiState.isSuccess) { showCompletePopup = uiState.isSuccess }
 
-    if (showPermissionPopup) {
-        CompletePopUp(
-            content = stringResource(id = R.string.check_family_permission_popup_content),
-            dismissText = stringResource(id = R.string.check_family_permission_popup_btn),
-            buttonColors = ButtonDefaults.buttonColors(containerColor = AppColors.purple2),
-            onDismissRequest = {
-                showPermissionPopup = false
-                navigateBack()
-            }
-        )
-    }
-    if (showCompletePopup) {
-        CompletePopUp(
-            content = stringResource(id = R.string.update_cycle_complete_popup_content),
-            dismissText = stringResource(id = R.string.update_cycle_complete_popup_btn),
-            buttonColors = ButtonDefaults.buttonColors(containerColor = AppColors.purple2),
-            onDismissRequest = {
-                showCompletePopup = false
-                navigateBack()
-            }
-        )
-    }
+    FamilyPermissionPopup(showPermissionPopup, navigateBack) { showCompletePopup = false }
+    FamilySettingCompletePopup(showCompletePopup, navigateBack) { showCompletePopup = false }
 
     UpdateCycleScreenUI(
         modifier = modifier,
