+++ app/src/main/res/values/strings.xml
@@ -207,6 +207,9 @@
     <string name="family_setting_label_leave_family">가족 탈퇴하기</string>
     <string name="family_setting_label_delete_family">가족 삭제하기</string>
 
+    <string name="family_setting_complete_popup_content">변경이 완료되었습니다.</string>
+    <string name="family_setting_complete_popup_btn">확인</string>
+
     <string name="family_invitation_link_title">가족 초대 링크</string>
     <string name="family_invitation_link_copy">초대 링크 복사</string>
     <string name="family_invitation_link_copied">링크가 복사되었습니다.</string>
@@ -240,8 +243,6 @@
     <string name="forgot_password_check_validation_warning">입력한 비밀번호와 일치하지 않습니다.</string>
 
     <string name="transfer_permission_title">권한 넘기기</string>
-    <string name="transfer_permission_complete_popup_content">변경이 완료되었습니다.</string>
-    <string name="transfer_permission_complete_popup_btn">확인</string>
 
     <string name="check_family_permission_popup_content">권한이 없습니다.</string>
     <string name="check_family_permission_popup_btn">확인</string>
@@ -291,7 +292,5 @@
 
     <string name="update_cycle_title">업로드 주기 변경</string>
     <string name="update_cycle_btn">업로드 주기 변경 완료</string>
-    <string name="update_cycle_complete_popup_content">변경이 완료되었습니다.</string>
-    <string name="update_cycle_complete_popup_btn">확인</string>
 
 </resources>
